# Prefix filter optimizations