@Cali0707 2023-09-27
/cc @pierDipi @creydr @matzew 

By the way, there is a behaviour change, however I think it is now more "correct". Before, if any of the attributes or values was an empty string, we would return that there was "NoFilter". However, I think that the more correct behaviour is to only return this "NoFilter" value if all of the attribute/value pairs have one or both be an empty string. WDYT? I can revert this behaviour change pretty easily if needed